requirements.txt
agate==1.9.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
attrs==24.2.0
babel==2.16.0
backoff==2.2.1
bcrypt==4.2.1
blinker==1.9.0
boto3==1.35.76
botocore==1.35.76
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
daff==1.3.46
dbt-adapters==1.10.4
dbt-common==1.14.0
dbt-core==1.9.0rc2
dbt-extractor==0.5.1
dbt-postgres==1.8.2
dbt-semantic-interfaces==0.7.4
deepdiff==7.0.1
Deprecated==1.2.15
distro==1.9.0
dnspython==2.7.0
duckdb==1.1.3
durationpy==0.9
email_validator==2.2.0
Faker==33.1.0
fastapi==0.115.6
filelock==3.16.1
flatbuffers==24.3.25
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
google-analytics-data==0.18.15
google-api-core==2.23.0
google-auth==2.36.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
greenlet==3.1.1
groq==0.13.0
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.26.5
humanfriendly==10.0
idna==3.10
importlib-metadata==6.11.0
importlib_resources==6.4.5
iniconfig==2.0.0
isodate==0.6.1
Jinja2==3.1.4
jiter==0.8.0
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kubernetes==31.0.0
leather==0.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mashumaro==3.14
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
mslex==1.3.0
narwhals==1.16.0
networkx==3.4.2
numpy==2.1.3
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.57.0
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-instrumentation-asgi==0.49b2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
opentelemetry-util-http==0.49b2
ordered-set==4.1.0
orjson==3.10.12
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parsedatetime==2.6
pathspec==0.12.1
pillow==11.0.0
plotly==5.24.1
pluggy==1.5.0
posthog==3.7.4
proto-plus==1.25.0
protobuf==5.29.1
psutil==6.1.1
psycopg2-binary==2.9.10
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pydeck==0.9.1
Pygments==2.18.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PySocks==1.7.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
s3transfer==0.10.4
selenium==4.27.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
snowplow-tracker==1.0.4
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
sqlparse==0.5.2
starlette==0.41.3
streamlit==1.40.2
streamlit-option-menu==0.4.0
sympy==1.13.3
taskipy==1.14.1
tenacity==9.0.0
text-unidecode==1.3
tokenizers==0.20.3
toml==0.10.2
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
trio==0.27.0
trio-websocket==0.11.1
typer==0.15.1
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
watchdog==6.0.0
watchfiles==1.0.0
websocket-client==1.8.0
websockets==14.1
wrapt==1.17.0
wsproto==1.2.0
zipp==3.21.0
